+++ src/main/java/com/l2jserver/gameserver/network/serverpackets/UserInfo.java
@@ -133,18 +133,18 @@ public final class UserInfo extends L2GameServerPacket
 		
 		writeD(_activeChar.getInventory().getTalismanSlots());
 		writeD(_activeChar.getInventory().canEquipCloak() ? 1 : 0);
-		writeD(_activeChar.getPAtk(null));
+		writeD((int) _activeChar.getPAtk(null));
 		writeD(_activeChar.getPAtkSpd());
-		writeD(_activeChar.getPDef(null));
+		writeD((int) _activeChar.getPDef(null));
 		writeD(_activeChar.getEvasionRate(null));
 		writeD(_activeChar.getAccuracy());
 		writeD(_activeChar.getCriticalHit(null, null));
-		writeD(_activeChar.getMAtk(null, null));
+		writeD((int) _activeChar.getMAtk(null, null));
 		
 		writeD(_activeChar.getMAtkSpd());
 		writeD(_activeChar.getPAtkSpd());
 		
-		writeD(_activeChar.getMDef(null, null));
+		writeD((int) _activeChar.getMDef(null, null));
 		
 		writeD(_activeChar.getPvpFlag());
 		writeD(_activeChar.getKarma());
